requirements.txt
google-cloud-secret-manager==2.16.1
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
langchain==0.0.292
langchain-community==0.0.6
youtube-transcript-api==0.6.1